GasStation: class with self-contained ensureGas, in its own file
Previous shape was custody-provider-specific: each provider duplicated
the same balance-check + send + poll-for-balance logic. None of it is
provider-specific (every provider sees the same chain through the same
ethers Provider/Signer pair on its gas wallet), so encapsulate it on the
GasStation itself.

- GasStation moves to its own file (gas-station.ts)
- Becomes a class taking (wallet, amount); exposes ensureGas(walletAddress)
- Drops ensureGas from CustodyProvider interface and from Fireblocks/DFNS
  provider implementations
- Delegates call this.custodyProvider.gasStation?.ensureGas(addr) directly

Custody surface is tighter — providers just supply a configured GasStation
when their env has gas-funding set, no behaviour to reimplement per provider.
