+++ src/services/direct/token-service.ts
@@ -75,8 +75,8 @@ export class DirectTokenService implements TokenService, EscrowService {
   }
 
   private async ensureGas(wallet: CustodyWallet): Promise<void> {
-    if (!this.custodyProvider.ensureGas) return;
-    await this.custodyProvider.ensureGas(await wallet.signer.getAddress());
+    if (!this.custodyProvider.gasStation) return;
+    await this.custodyProvider.gasStation.ensureGas(await wallet.signer.getAddress());
   }
 
   async createAsset(
